# docs/powerusyservice/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a backend web service created for Powerusy operations.",
        "title": "Powerusy backend service",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Alao Ramon, Nabil Abubakar",
            "url": "http://www.swagger.io/support",
            "email": "alao.adebusy@gmail.com, nabbo247@yahoo.com"
        },
        "license": {
            "name": "AddyTech Solution Ltd",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8060",
    "basePath": "/",
    "paths": {
...
        "/api/shippers/ApproveShippersDocument": {
            "patch": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shipper"
                ],
                "summary": "Approve shippers document",
                "parameters": [
                    {
                        "description": "User Details",
                        "name": "shipperDoc",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ShipperApprovalReqIn"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseMessage"
                        }
                    }
                }
            }
        },
...
        "/api/shippers/UploadShippersDocument": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shipper"
                ],
                "summary": "Upload shippers document",
                "parameters": [
                    {
                        "description": "User Details",
                        "name": "shipperDoc",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ShipperIn"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseMessage"
                        }
                    }
                }
            }
        },
...
        "/api/users/CreateNewUser": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Creates new user",
                "parameters": [
                    {
                        "description": "User Details",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UserIn"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.UserOut"
                        }
                    }
                }
            }
        },
...
        "/api/users/Login": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "log user into the syste",
                "parameters": [
                    {
                        "description": "Login user",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.LoginIn"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.UserOut"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.LoginIn": {
            "type": "object",
            "properties": {
                "Password": {
                    "type": "string"
                },
                "Username": {
                    "type": "string"
                }
            }
        },
        "models.ResponseMessage": {
            "type": "object",
            "properties": {
                "responseCode": {
                    "type": "string"
                },
                "responseMessage": {
                    "type": "string"
                }
            }
        },
        "models.ShipperApprovalReqIn": {
            "type": "object",
            "properties": {
                "ApprovedBy": {
                    "type": "integer"
                },
                "Comment": {
                    "type": "string"
                },
                "CompanyName": {
                    "type": "string"
                },
                "Statusid": {
                    "type": "integer"
                },
                "UserId": {
                    "type": "integer"
                }
            }
        },
        "models.ShipperIn": {
            "type": "object",
            "properties": {
                "CompanyName": {
                    "type": "string"
                },
                "Location": {
                    "type": "string"
                },
                "Phonenumber": {
                    "type": "string"
                },
                "UserId": {
                    "type": "integer"
                },
                "tin": {
                    "type": "string"
                },
                "tinpassword": {
                    "type": "string"
                }
            }
        },
...
        "models.UserIn": {
            "type": "object",
            "properties": {
                "Email": {
                    "type": "string"
                },
                "Firstname": {
                    "type": "string"
                },
                "Lastname": {
                    "type": "string"
                },
                "Middlename": {
                    "type": "string"
                },
                "Password": {
                    "type": "string"
                },
                "Phonenumber": {
                    "type": "string"
                },
                "Roleid": {
                    "type": "integer"
                },
                "Username": {
                    "type": "string"
                }
            }
        },
        "models.UserOut": {
            "type": "object",
            "properties": {
                "Accountnumber": {
                    "type": "string"
                },
                "Approvedby": {
                    "type": "string"
                },
                "Dateapproved": {
                    "type": "string"
                },
                "Statusid": {
                    "type": "string"
                },
                "authcode": {
                    "type": "string"
                },
                "dateadded": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "firstname": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lastname": {
                    "type": "string"
                },
                "middlename": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phonenumber": {
                    "type": "string"
                },
                "roleid": {
                    "type": "integer"
                },
                "status": {
                    "type": "boolean"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}